requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.5
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
packaging==21.0
pluggy==0.13.1
prompt-toolkit==3.0.19
psycopg2==2.9.1
py==1.10.0
pylibmc==1.6.1
pymemcache==3.5.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
